+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - added `set_settings`
 - added `get_reviews`
 - added Shopgate `custom_fields` support to customer & customer addresses (see Readme)
+- added support for categories with dynamic product groups
 - fixed customer address type mapping in case billing and shipping address are the same
+- fixed product sort to use parent ID instead of child ID for sort order positions in a category
 
 # 1.6.4